requirements.txt
asgiref==3.3.1
Django==3.1.5
django-extensions==3.1.0
django-model-utils==4.1.1
django-money==1.3
django-phonenumber-field==5.0.0
phonenumberslite==8.12.17
py-moneyed==0.8.0
pytz==2020.5
sqlparse==0.4.1
